update customer issue

+++ app/Http/Controllers/CustomerController.php
@@ -40,7 +40,7 @@ class CustomerController extends Controller
         $data['schedules'] = $gameModel->prepareGameData($id);
         $data['slot_time_id'] = $time_id;
 
-        $closeMinutes = CloseTime::pluck('minutes')->first();
+        $closeMinutes = (int) CloseTime::pluck('minutes')->first();
         $data['close_time'] = $closeMinutes;
         $slots = CreateGameScheduleModel::with('digitMaster', 'providerSlot')
             ->where('betting_providers_id', $id)
@@ -192,11 +192,9 @@ class CustomerController extends Controller
                 'user' => $user,
                 'transactions' => $transactions
             ]);
-        } 
-        catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
+        } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             return redirect()->back()->with('error', 'User not found');
-        } 
-        catch (\Exception $e) {
+        } catch (\Exception $e) {
             Log::error('Payment history error: ' . $e->getMessage());
             return redirect()->back()->with('error', 'Failed to retrieve payment history');
         }
